requirements.txt
certifi
selenium==4.2.0
pandas
python-dotenv
chromedriver-autoinstaller
geckodriver-autoinstaller
urllib3
pyautogui
bs4
requests
snscrape
pandas
strsimpy
mysql.connector
nltk
spacy>=3.4.4
jellyfish
scikit-learn==1.2.0
mysql-connector-python
sqlalchemy
matplotlib
vaderSentiment
